+++ shuup/admin/static_src/picotable/picotable.js
@@ -530,7 +530,7 @@ const Picotable = (function (m, storage) {
             }
             return m("tr", rowSettings, Util.map(data.columns, function (col, idx) {
                 var content;
-                if (idx === 0 && massActions.length) {
+                if (idx === 0 && massActions.length && (!item.hasOwnProperty("popup") || item.popup === false)) {
                     content = m("div.input-checkbox", { onclick: preventSelect }, [
                         m("input[type=checkbox]", {
                             id: item._id,
@@ -545,8 +545,17 @@ const Picotable = (function (m, storage) {
                             title: gettext("Edit")
                         }, m("i.fa.fa-edit")) : null),
                     ]);
-                }
-                else {
+                } else if (idx === 0 && massActions.length && item.popup === true) {
+                    content = m("div.input-checkbox", { onclick: preventSelect }, [
+                        m("button[type=button]", {
+                            class: "browse-btn btn btn-primary btn-sm",
+                            onclick: Util.boundPartial(ctrl, ctrl.pickObject, item)
+                        }, [
+                            m("i.fa.fa-folder"),
+                            gettext(" Select")
+                        ])
+                    ]);
+                } else {
                     content = item[col.id] || "";
                 }
                 if (col.raw) {
@@ -673,7 +682,7 @@ const Picotable = (function (m, storage) {
                         rowClasses.push(...item._extra.class.split(" "));
                     }
                     return m("." + rowClasses.join("."), [
-                        (line.class && massActions.length ?
+                        (line.class && massActions.length && (!item.hasOwnProperty("popup") || item.popup === false) ?
                             m("div.input-checkbox", { onclick: preventSelect }, [
                                 m("input[type=checkbox]", {
                                     id: item._id,
@@ -685,7 +694,17 @@ const Picotable = (function (m, storage) {
                                 m("label", { for: item._id, }),
                                 (item._url ? m("a.edit", {href: item._url}, m("i.fa.fa-edit")) : null)
                             ])
-                            : null
+                        : (line.class && massActions.length && item.popup === true ?
+                                m("div.input-checkbox", { onclick: preventSelect }, [
+                                    m("button[type=button]", {
+                                        class: "browse-btn btn btn-primary btn-sm",
+                                        onclick: Util.boundPartial(ctrl, ctrl.pickObject, item)
+                                    }, [
+                                        m("i.fa.fa-folder"),
+                                        gettext(" Select")
+                                    ])
+                                ])
+                            : null)
                         ),
                         (line.title ? m(".col.title", line.title) : null),
                         m(".col.value", line.text)